Temperature torch.float64

+++ clip.py
@@ -60,7 +60,7 @@ class CLIP(nn.Module):
 
         # "The learnable temp parameter was initialized to the equivalent of 0.07."
         # self.temp = nn.Parameter(torch.tensor((0.07,)))
-        self.temp = nn.Parameter(torch.tensor((1,)))
+        self.temp = nn.Parameter(torch.tensor((1,), dtype=torch.float64))
 
         # self.gt = torch.arange(batch_size)
